requirements.txt
certifi==2020.6.20
chardet==3.0.4
Click==7.0
idna==2.6
mss==3.3.1
Pillow==7.2.0
pyperclip==1.7.0
pytest-runner==5.2
requests==2.20.0
urllib3==1.22
